Change actions to fire at release, not press
This works around what I suspect is either a sway or GTK bug.  When the
on-click action activates a window on a different workspace, the mouse is
immediately warped to that workspace.  This apparently causes waybar to never
get the mouse-up event, resulting in later interactions with the bar to work
incorrectly until it gets a corresponding mouse-up event.

Triggering on button release fixes the problem.